# docs/openapi/backend.json
{
  "components": {
    "schemas": {
...
      "ColumnAttributeExplorationDetails": {
        "additionalProperties": true,
        "description": "A ColumnAttribute's own owning Term and every linked Column, for\nexpanding it on the graph.\n\n`term` is `None` only when the attribute has no owning Term (or it's\nout of scope). `columns` is one ordered page of *every* Column\nHAS_ATTRIBUTE/SEMANTIC_FK-linked to this attribute \u2014 not just the\n\"primary\" one whichever expansion grafted this attribute on already\nknew about \u2014 since a shared attribute (e.g. a common `user_id`-shaped\none) can be linked from many Columns across many Tables at once. See\n`fetch_column_attribute_exploration_details`.",
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnAttributeExplorationColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "columns_total": {
            "default": 0,
            "title": "Columns Total",
            "type": "integer"
          },
          "term": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TermSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ColumnAttributeExplorationDetails",
        "type": "object"
      },
...
      "ColumnExplorationDetails": {
        "additionalProperties": true,
        "description": "A Column's own ColumnAttribute, outgoing FOREIGN_KEY target Column,\nincoming FOREIGN_KEY source Columns, and referencing Sql queries, for\nexpanding it on the graph.\n\n`column_attribute`/`foreign_key_column` are `None` for a column with no\nColumnAttribute/no outgoing FK; `referencing_columns` is empty when no\nother Column's own FK points at this one (the reverse of\n`foreign_key_column`); `sql_queries` is empty for a column no stored\nquery ever referenced directly \u2014 see `fetch_column_exploration_details`.",
        "properties": {
          "column_attribute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColumnExplorationAttribute"
              },
              {
                "type": "null"
              }
            ]
          },
          "foreign_key_column": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColumnExplorationForeignKeyColumn"
              },
              {
                "type": "null"
              }
            ]
          },
          "referencing_columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnExplorationForeignKeyColumn"
            },
            "title": "Referencing Columns",
            "type": "array"
          },
          "sql_queries": {
            "items": {
              "$ref": "#/components/schemas/TableSqlQuery"
            },
            "title": "Sql Queries",
            "type": "array"
          }
        },
        "title": "ColumnExplorationDetails",
        "type": "object"
      },
...
      "DataExplorationGraph": {
        "additionalProperties": true,
        "description": "``{nodes, links}`` for the data layer, in one response.",
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/ExplorationEdge"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/DataGraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "DataExplorationGraph",
        "type": "object"
      },
...
      "DataResponse_ColumnAttributeExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ColumnAttributeExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[ColumnAttributeExplorationDetails]",
        "type": "object"
      },
      "DataResponse_ColumnExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ColumnExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[ColumnExplorationDetails]",
        "type": "object"
      },
...
      "DataResponse_DataExplorationGraph_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataExplorationGraph"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[DataExplorationGraph]",
        "type": "object"
      },
...
      "DataResponse_ExplorationLinkPath_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExplorationLinkPath"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[ExplorationLinkPath]",
        "type": "object"
      },
      "DataResponse_ExplorationRelatedNodes_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExplorationRelatedNodes"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[ExplorationRelatedNodes]",
        "type": "object"
      },
...
      "DataResponse_SemanticExplorationGraph_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SemanticExplorationGraph"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[SemanticExplorationGraph]",
        "type": "object"
      },
      "DataResponse_SqlAttributeExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SqlAttributeExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[SqlAttributeExplorationDetails]",
        "type": "object"
      },
...
      "DataResponse_SqlExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SqlExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[SqlExplorationDetails]",
        "type": "object"
      },
...
      "DataResponse_TableExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TableExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[TableExplorationDetails]",
        "type": "object"
      },
...
      "DataResponse_TermExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TermExplorationDetails"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[TermExplorationDetails]",
        "type": "object"
      },
...
      "ExplorationRelatedNodes": {
        "additionalProperties": true,
        "description": "One page of related nodes; ``total`` counts every related node.",
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ExplorationRelatedNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "title": "ExplorationRelatedNodes",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Auto Ontology API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/exploration/column-attributes/{attr_id}/details": {
      "get": {
        "description": "The ColumnAttribute's own owning Term and one ordered page of every\nlinked Column, for expanding it on the Exploration graph.",
        "operationId": "get_column_attribute_exploration_details_api_exploration_column_attributes__attr_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "attr_id",
            "required": true,
            "schema": {
              "title": "Attr Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ColumnAttributeExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Column Attribute Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/columns/{column_id}/details": {
      "get": {
        "description": "The Column's own ColumnAttribute, FK target, referencing FK Columns and Sql queries, for expanding it on the Exploration graph.",
        "operationId": "get_column_exploration_details_api_exploration_columns__column_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "column_id",
            "required": true,
            "schema": {
              "title": "Column Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ColumnExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Column Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
...
    "/api/exploration/graph": {
      "get": {
        "description": "Return the full data-layer Exploration graph (``{nodes, links}``).\n\nLets the client render the data graph from one request instead of\nwalking the catalog tree (databases \u2192 schemas \u2192 tables) with a request\nper level. Node count is capped at ``MAX_EXPLORATION_GRAPH_NODES``\nregardless of *limit*.",
        "operationId": "get_data_exploration_graph_api_exploration_graph_get",
        "parameters": [
          {
            "description": "Max nodes returned; always capped at 500 server-side.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "description": "Max nodes returned; always capped at 500 server-side.",
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_DataExplorationGraph_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Data Exploration Graph",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/nodes/{node_id}/relationships": {
      "get": {
        "description": "Return one ordered page of nodes related to ``node_id``.",
        "operationId": "get_exploration_node_relationships_api_exploration_nodes__node_id__relationships_get",
        "parameters": [
          {
            "description": "Table id when *layer* is ``data``, Term id when it is ``semantic``.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "description": "Table id when *layer* is ``data``, Term id when it is ``semantic``.",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "description": "Which graph to walk: ``data`` relates tables sharing a SQL query or a foreign key, ``semantic`` relates terms sharing a table.",
            "in": "query",
            "name": "layer",
            "required": true,
            "schema": {
              "description": "Which graph to walk: ``data`` relates tables sharing a SQL query or a foreign key, ``semantic`` relates terms sharing a table.",
              "enum": [
                "data",
                "semantic"
              ],
              "title": "Layer",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ExplorationRelatedNodes_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exploration Node Relationships",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/semantic-graph": {
      "get": {
        "description": "Return the full semantic-layer Exploration graph (``{nodes, links}``).\n\nLets the client render the semantic graph from one request instead of\nfetching related terms once per term (an N+1). Node count is capped at\n``MAX_EXPLORATION_GRAPH_NODES`` regardless of *limit*.",
        "operationId": "get_semantic_exploration_graph_api_exploration_semantic_graph_get",
        "parameters": [
          {
            "description": "Max nodes returned; always capped at 500 server-side.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "description": "Max nodes returned; always capped at 500 server-side.",
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_SemanticExplorationGraph_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Semantic Exploration Graph",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/sql-attributes/{attr_id}/details": {
      "get": {
        "description": "A SqlAttribute's own Sql query and owning Term, for expanding it on the Exploration graph.",
        "operationId": "get_sql_attribute_exploration_details_api_exploration_sql_attributes__attr_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "attr_id",
            "required": true,
            "schema": {
              "title": "Attr Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_SqlAttributeExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sql Attribute Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/sql/{sql_id}/details": {
      "get": {
        "description": "CustomAnalysis, Column, Table and SqlAttribute nodes hanging off this Sql node, for expanding it on the Exploration graph.",
        "operationId": "get_sql_exploration_details_api_exploration_sql__sql_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "sql_id",
            "required": true,
            "schema": {
              "title": "Sql Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_SqlExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sql Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
...
    "/api/exploration/tables/{table_id}/details": {
      "get": {
        "description": "Columns-adjacent SQL and one ordered page of Term details for an Exploration modal.",
        "operationId": "get_table_exploration_details_api_exploration_tables__table_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TableExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Table Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/terms/{term_id}/details": {
      "get": {
        "description": "One ordered page of Tables linked to a Term, for expanding it on the Exploration graph.",
        "operationId": "get_term_exploration_details_api_exploration_terms__term_id__details_get",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TermExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Term Exploration Details",
        "tags": [
          "exploration"
        ]
      }
    },
    "/api/exploration/terms/{term_id}/path/{other_term_id}": {
      "get": {
        "description": "The real hop chain(s) behind one term\u2194term Exploration graph edge.\n\nLets the client graft the actual Table (and, for a HAS_ATTRIBUTE/\nSEMANTIC_FK hop, Column/ColumnAttribute) nodes on and highlight them\nonce the edge between ``term_id``/``other_term_id`` is clicked, instead\nof just showing the collapsed `GraphLink.relationship_types` label.",
        "operationId": "get_semantic_link_path_api_exploration_terms__term_id__path__other_term_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "other_term_id",
            "required": true,
            "schema": {
              "title": "Other Term Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ExplorationLinkPath_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Semantic Link Path",
        "tags": [
          "exploration"
        ]
      }
    },
...
  }
}